Programmatic restarting of USB camera device. Important: requires installation of WDK, addition of devcon.exe directory to "Path" system variable, and the "Restart Cameras.vi" to be exported (via LabVIEW VI server prefs. on Camera Server), and "Restart Cameras.vi" to be added to the Startup programs folder "C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Startup".
On occasion, a camera would appear present but not communicate (timeout). The solution to this was either to cycle the USB connector, or to disable and then enable the device in Device Manager. This code performs the latter, and may be called whenever an image acquisition error is returned (and indeed the image acquisition can then be attempted once again). Admittedly not the most elegant implementation, but it is not possible to run cmd.exe with Admin privilege from LabView, hence the shortcut to a batch file with the "Run as Administrator" property set! Editing the batch file does require the user to manually extract Hardware IDs from the device manager for each camera.... The batch file is commented to help out (or remind us what to do in the distant future!!) in this regard.