Minor changes to: - Email_With_Log.vi: Now uses new authenticated SMTP system for sending emails. - SMTP_email_MultiAttach_Authenticated.vi: Wired up Username, Password, and Reply to Address inputs.
Edits to Check_Server.vi:
CPP’s new authenticated email system requires that the server port address be included in the server URL. The URL input string is now parsed to separate the port number from the rest of the URL. If a port number is found in the URL, then TCP Open Connection will use that port number. Otherwise, it will use port 25.